Separates logic into BootScene and GameScene.
